+++ packages/ui-client/src/providers/index.ts
@@ -1,1 +1,2 @@
 export * from './ModalProvider';
+export { default as TooltipProvider } from './TooltipProvider';
